requirements/dev.txt
# Development requirements
# Includes base requirements plus development tools

-r base.txt

# Development Server
# Django runserver is included with base Django

# Code Quality & Linting
ruff>=0.1.6
black>=23.11.0
isort>=5.12.0
pre-commit>=3.5.0

# Type Checking
mypy>=1.7.0
django-stubs[compatible-mypy]>=4.2.0
djangorestframework-stubs[compatible-mypy]>=3.14.0

# Security Scanning
bandit>=1.7.5
pip-audit>=2.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0
freezegun>=1.2.0

# Development Tools
django-extensions>=3.2.0  # shell_plus, graph_models, etc.
ipython>=8.0.0  # Better shell
ipdb>=0.13.0  # Debugger

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0